chore: move tests to a new folder
